+++ src/stac_asset/constants.py
@@ -1,2 +1,0 @@
-DEFAULT_ASSET_FILE_NAME = "asset.json"
-"""When downloading an asset, the default file name for the asset itself."""